hostapd: option 'pbc_in_m1' for war PBC with WIN7
WPS capability discovery workaround for PBC with Windows 7
Windows 7 uses incorrect way of figuring out AP's WPS capabilities by acting
as a Registrar and using M1 from the AP. The config methods attribute in that
message is supposed to indicate only the configuration method supported by
the AP in Enrollee role, i.e., to add an external Registrar. For that case,
PBC shall not be used and as such, the PushButton config method is removed
from M1 by default. If pbc_in_m1=1 is included in the configuration file,
the PushButton config method is left in M1 (if included in config_methods
parameter) to allow Windows 7 to use PBC instead of PIN (e.g., from a label
in the AP) (Reference hostapd.conf)

Add option 'pbc_in_m1' to enable this workaround from UCI.

Change-Id: Id5a20032d3c759fb6b44de8c4c8081db9f5b2b31
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>