[base-files] Add generic debug mechanism
The goal of this mechanism is to provide an easy and decentralized way
for a package to register a set of debugging commands, in the case where
someone wants to gather all the logs in the system and store them for
later analysis.

The applications can register a set of debug commands by adding a
script in /lib/debug. The "sysdebug" command will recursively execute
all these scripts, log their output and format it accordingly.

Change-Id: If38cd771d6ca7dcbc81b7c3071867e1f28cd62b0
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>