[ar71xx] Remove kmod-art2 from wireless profile
This takes up lot of space and the images are not created
with the error "root.squashfs-64k is too big".

Remove it for now and if required add this package back by
figuring out few other utility packages which is of
not high significance for the minimal wireless profile.

Change-Id: I560ce64bfbc4d1de4aab02185fba3d21d598a19d
Signed-off-by: Mohammmed Shafi Shajakhan <mohammed@codeaurora.org>