scripts: add ubinize-image.sh to generate ubi images
This is a wrapper for the ubinize tool which integrates auto-generation
of the ubinize.cfg for common volume layouts with and without a kernel
volume.
It supports auto-detecting the rootfs-type and decides whether or not
to include a rootfs_data volume based on whether the rootfs is ubifs
or not (and thus is read-only and needs rootfs_data ubifs overlay).
The script allows to create layouts as found in all current
ubinize{,-overlay}.cfg files using

ubinize-image.sh --no-kernel root.{ubifs,squashfs} output.ubi $UBI_OPTS

It also includes support for adding ubootenv and ubootenv2 volumes
typically used by U-Boot for storing its environment in UBI using the
"--uboot-env" parameter.
See also
https://gitorious.org/openwrt-oxnas/openwrt-oxnas/source/target/linux/
oxnas/image/Makefile

Signed-off-by: Daniel Golle <daniel@makrotopia.org>

 create mode 100755 scripts/ubinize-image.sh

Change-Id: Ie20c5d405981f98f03fcd7371b4504105c6f0a78
git-svn-id: svn://svn.openwrt.org/openwrt/trunk@41121 3c298f89-4303-0410-b956-a3cf2f4a3e73
Signed-off-by: Nisar Ahamed Khan <nkhan@codeaurora.org>