[qca-nss-ecm] Added required iptables-mod-extra
ECM uses a -m physdev rule in it's additions to firewall.  This
requires adding iptables-mod-extra.

As an additional note, we should definitely look carefully at this.  It
seems to me that any firewall rule reordering could break the usage of
the firewall rules.  For example, if there were a -j DROP rule before
this, it could make the FW rule ineffective.

Change-Id: Ibc5fe67843cc5fd05c0ad10d77a1803d4d01216d
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>