Use "package" as an XML key rather than package name
This change is required as some packages starts with a number, which
(when enabled) results in generating an XML file which cannot be
processed using the standard XML perl libraries. These do not support
having a number as a first character of an XML tag.

So we'll change the format from:
        <pkgA build_dir="..." local_src="..."/>
        <pkgB build_dir="..." local_src="..."/>
        <pkgC build_dir="..." local_src="..."/>
        ...
To this:
        <package name="pkgA" build_dir="..." local_src="..."/>
        <package name="pkgB" build_dir="..." local_src="..."/>
        <package name="pkgC" build_dir="..." local_src="..."/>
        ...

Change-Id: I1ca71f58aab3825c8d3601bac0d9b2dc90ad840f
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>