[ar71xx] Error-out if the firmware images are too big
By default, the OpenWrt code only prints a warning if the generated
image gets too big to fit in flash. However, for our internal usage,
we expect the boards within a single profiles to have enough memory
to include all the packages, else it's an error.

So we'll change this behavior and error-out if the size gets bigger
than the allocated size on the platform. This will allow us to catch
problematic flash usage increase from the internal components earlier
in the integration process.

Change-Id: I5a4d2aebf4619af31e3c6d8296ee2f01ce9c0cf0
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>