iwinfo: Add qca-backports support
By default, iwinfo will only be built with nl80211 support if the
mac80211 driver package is enabled. As QSDK is using an alternate package
called qca-mac80211, which is equivalent to mac80211, but synced at a
different point in time, we need iwinfo to be built with nl80211 support
when that package is enabled as well.

Change-Id: Iaac3916ddee2934d036fc9e15c82c0a5088cd97d
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>