[qca-nss-drv] Add uci option to support smp affinity
The UCI configuration "nss.global.smp_affinity" can now be used
to set/unset affinity. The default value will be set to whatever
the driver is using, and it may be overriden with specified in the
config.

This change also removes all the qca-nss-connmgr files which are
no longer used since they moved into the ECM package.

Change-Id: I0e1cdfc29d105f0421a7503f50eb1ac46ae09b99
Signed-off-by: Mohammed Shafi Shajakhan <mohammed@codeaurora.org>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>