[generic] Add an OpenWrt menuconfig option to enable KMEMLEAK
Kernel memory leak can be useful to debug memory leak. This change will
add an OpenWrt menuconfig option to enable/disable it (disable by
default).

This is also adding the default OpenWrt value that will be statically
set when enabling KMEMLEAK. This is required in order to avoid prompting
for user input when enabling the option.

Change-Id: Ie195be03116d40f251096c359f965d8c8d6a04b6
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>