netfilter: nf_ct_sip: don't drop packets with offsets pointing outside the packet
Some Cisco phones create huge messages that are spread over multiple packets.
After calculating the offset of the SIP body, it is validated to be within
the packet and the packet is dropped otherwise. This breaks operation of
these phones. Since connection tracking is supposed to be passive, just let
those packets pass unmodified and untracked.
Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
Cc: William Roberts <bill.c.roberts@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Abirama Sundari Sivakumar <asivakum@codeaurora.org>

Change-Id: Ib69a697a4553cfc8bb8d04d99474d6d615180e7a
Signed-off-by: Abirama Sundari Sivakumar <asivakum@codeaurora.org>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>