Revert "[qca-nss-ecm] Fix installing the service by default"
This mechanism is there so the qca-nss-ecm-noload variant can be
installed, disabled by default at init, and can be enabled at run-time
with the "enable" command.

This change causes the -noload variant to load ECM, which is not wanted
as the corresponding sysctl values will not be set. Therefore causing
the packets to not be forwarded through the bridge interface.

This reverts commit 1507429ff1d59fe9f543f1e16bd0f9ee7ff7526b

Change-Id: I1e4785ac6ed3de88d68376ae3b0bd2845aabee51
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>