[ipq806x] Build ext4 support into the kenrel
ext4 was built as a kernel module, but as we want to enable booting
some platforms from a block device, we will want to mount the
rootfs_data from an ext4 partition, so we need to get it supported
at the preinit stage.
As preinit happens before loading the modules, we need to add support
for ext4 into the kernel.

Change-Id: Icaf2896d3c87bd99f5ed24a4c6d7b57e1d063b04
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>