[base-files] Simplify ubifs mount implementation
We want to extend the boot support to eMMC, and we need to way to add
new device type reliably.
Current implementation checks whether a ubi_rootfs_data partition is
present before mounting jffs2 (doesn't mount it if it finds one), which
is difficult to maintain. Instead, this is changing it for the approach
of checking whether the FS is present without mounting it. It shouldn't
have any functional impact, but will allow us to add more devices
without having to modify the entire mount scripts set for every new
device type we add.

Change-Id: Idd00d0021e45fc099c613f65630657729672ba1a
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>