[package] kernel/modules: change the pppoe and pppox module load order.
After the re-work and clean-up of kernel changes for PPPoE
acceleration NSS driver uses these modules and their load order
must be less than the NSS driver.

qca-nss-drv's autoload order number is 32. In this change, pppox and pppoe are
selected as 31 and 32 respectively. In alphabeical order, pppoe becomes before
the qca-nss-drv and is loaded before it.

Change-Id: Iccc3809d77fcc193119cb293cb800b8debeb3b66
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>