netifd: Set source IP for DHCP static routes as well
Commit ce92f6650bd8a86db04c7a6cbb58e7fdb200a7e6 added source IP support
for DHCP default routes. As a side effect of this change the default route
could be present twice in netifd (once with source IP set and once with
source IP unset) if it was sent by the server in both the router and static
route options. Therefore add source IP support as well for static routes as this
case was not considered. Additional remove unused parameter type.

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@43645 3c298f89-4303-0410-b956-a3cf2f4a3e73

Change-Id: Id1689461c8c9eef5fc16232f081380e7fe992514
Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>
Signed-off-by: Parameshwaran Kanagaraj <pkanag@codeaurora.org>