ncurses: move tinfo in a separate libraries
Some packages (such as readline) build without ncurses but still need
the tinfo library provided by this package.
In order to be space efficient, we'll build the tinfo library as a
separate object, and provide it in a standalone package. This will
allow other application to depend on this package without needing the
entire ncurses library.

Change-Id: I033581e6753e768bb02661a36376ca95e63775da
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>