build: allow openwrt.git packages to be replaced by feeds
Currently, replacing a package available in openwrt.git requires
modifications in openwrt.git, or requires duplicating the package in a
feed but with a different name, which causes all kind of problems
related to dependencies (all packages selecting it would have to be
modified accordingly to select the new package).

With this change, if a package with the same name is present both in
feeds/ and package/ folders, the one in package/ will be ignored,
the feeds taking precedence over the version in package/, both in the
menuconfig and during the build.

This mechanism is particularly useful for vendor tree, or in general for
application which needs to replace one particular package which exists
within openwrt.git by a newer version.

Note that priorities between feeds is inherently provided by
./scripts/feeds, which parses feeds.conf line by line and installs only
the first package it found. The higher the feed is in feeds.conf, the
higher priotity its packages will have.

Obviously, this gives lowest priority to packages in openwrt.git.

Change-Id: Ie73c550b5b41bbeefa0674e1ff5a5de1f405d7ad
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>