build: override package from feeds
Feeds script will ignore the package if it's already installed in
package/. So we'll add a special case here, which will take check
whether the package is already installed from the core openwrt pkgs
or if it's installed from a feed. If installed from the core packages
it will be overriden.

Change-Id: I7462ce2f18fd1e2114bda61046363edbc399c45f
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>