[qca ecm] Use modules.d to load ECM
ECM depends on listening to netdevice notifier messages to pick up
some information on system state.  Currently, ECM is loaded in an
rc.d script after all modules are loaded.

This changes the location in which ECM is loaded to modules.d rather
than rc.d, making it early enough to pick up all changes that occur in
networking during init.

This means that any system configuration that needs to be in place for
ECM to function may not be configured for the time period between module
load and the init.d script running.  This is considered acceptable.

Change-Id: I9a33aee31a78de1b843f19116ac7ff0fbd091143
Signed-off-by: Tushar Mathur <tushar@codeaurora.org>