ipq806x: copy the correct ELF file to bindir
The kernel is built twice ones for normal boot and once
for initramfs. We were copying the initramfs version
into the bindir and most people run the normal version
of the kernel, so they got a few mismatched symbols.
Namely the build time changed and it effect a few
variables in the image.

Change-Id: Ic697a4450fb3e7f6c7ef7ecf86ba5c9c220198b8
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>