build: propagate override information to .config file
With this change, override information is now parsed from the metadata
and put in the %packages hash. A new hash - %overrides - is created and
exported, to be used during the .config-package.in generation.

If an override is detected, a new option CONFIG_OVERRIDE_PKGS will be
created in the .config, and will contain a space-separated list of all
the overridden packages.

Change-Id: I29cb191febf45ad4026ad6d34fb8920348d19766
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>