toolchain: Add hardening options and make them the default
This patch adds security hardening options to uClibc and all OpenWRT
packages. This includes stack protection and RELRO. Stack protection is
requierd to protect against trivial buffer overflow attacks. RELRO is
also considered necessary precaution to protect the PLT (Procedure
Linkage Table) and GOT (Global Offset Table). Combining these
mitigations can protect against a large range of these class of attacks.
This change will make sure that bugs that may be present in the code do
not end up as security vulnerabilities.
As these options impact the whole build with all pacakges, We're also
adding a build option (in DEVEL) to disable these from the menuconfig.
That said, we make the disable mechanism off by default as the impact
of these should be minimal, and it's a great security enhancement.

Change-Id: Ic29733d01405dcb6bb179a1d249d605075ee42bc
Signed-off-by: Nathan Crandall <nathanc@codeaurora.org>