[qca-nss-ecm] Add iptables-mod-conntrack-extra as a dependency
ECM makes use of the "mark" field in nf_conn. Therefore it needs the
option CONFIG_NF_CONNTRACK_MARK enabled in the kernel to compile
properly.

This option is enabled through other means in most profiles, but we want
to make sure it's enabled all the time once ECM gets selected. So we'll
automatically select iptables-mod-conntrack-extra when selecting ECM.
This package enables the kernel option automatically.

Change-Id: Icd7a8aad5164b5821170bfe7c25ded1190fc9b3c
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>