hostapd: Make IPv6 NA multicast-to-unicast conversion configurable
This can be used with Proxy ARP to allow multicast NAs to be
forwarded to associated STAs using link layer unicast delivery.
This used to be hardcoded to be enabled, but it is now disabled by
default and can be enabled with na_mcast_to_ucast=1. This functionality
may not be desired in all networks and most cases work without it, so
the new default-to-disabled is more appropriate.

This patch is cherry-picked from upstream hostapd git.
http://w1.fi/cgit/hostap/commit/?id=4a7ce984690d373484fcef388b468e034e448c92

Change-Id: I2150a10b209b2551b126b18ef8eb8443940082eb
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>