[ipq806x] setmac: Modifying setmac scripts to write in mtdblock
As we were not able to write to mtdblock, we made setmac script
to write in mtd in commit 1998a843ca1897025b3b46edbe665c10ae9e663b.
Now changes has been done to driver, so we are able to write in
mtdblock, thus modifying setmac scripts to write in mtdblock.

Change-Id: Ie69a829c546aa346712c773bf9bcd227b03fd0d6
Signed-off-by: Anto Norbert <norbrt@codeaurora.org>