[ipq40xx] Build ext4 support into the kernel
ext4 was built as a kernel module, but as we want to enable booting
some platforms from a block device, we will want to mount the
rootfs_data from an ext4 partition, so we need to get it supported
at the preinit stage.
As preinit happens before loading the modules, we need to add support
for ext4 into the kernel.

Change-Id: I8e8aea1210cf17dbc2a4a17b1f3a93b065a4f9aa
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>