[uClibc] Avoid interactive queries during builds
Build options were missing from the .config when SSP was enabled. So
we'll add their default value to the common config available in the
uClibc package, and we'll modify the UCLIBC_HAS_SSP match in common.mk
to avoid matching UCLIBC_HAS_SSP_COMPAT as well.

This should ensure smooth builds with SSP enabled in the global config
as well.

Change-Id: Ie37df161750c0bcbd1ff06837f02983fd6feeadc
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>