include/scan.mk: Add -a argument to the grep to treat all Makefiles as a text file.
Hello,

I found out that in some rare cases grep can treat Makefile as a binary file. That happened to me on UTF-8 Gentoo if Makefile contained a character which was from extended ASCII table.
Without this patch the output of the grep command in the $(FILELIST) target is not valid (contains line "Binary file matches") and following commands fail to create feed index file.

Best Regards,
Martin Strbacka

Signed-off-by: Martin Strbacka <martin.strbacka@nic.cz>

Change-Id: I37d0dac5df893816608ae68e155d9baf4afd57f2
git-svn-id: svn://svn.openwrt.org/openwrt/trunk@45966 3c298f89-4303-0410-b956-a3cf2f4a3e73
Signed-off-by: Srinivas dodda <doddas@codeaurora.org>