gcc/musl: rework SSP-support
Make musl provide libssp_nonshared.a and make GCC link it unconditionally
if musl is used. This should be a no-op if SSP is disabled and seems to be
the only reliable way of dealing with SSP over all packages due to the mess
that is linkerflags handling in packages.

Signed-off-by: Steven Barth <steven@midlink.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@46108 3c298f89-4303-0410-b956-a3cf2f4a3e73
(cherry picked from commit a833be775656312fe9a1fb5b3d235d03dc141533)

Change-Id: I37f5c5f492959056b3c1c841f7b315f4fa308517
Signed-off-by: Anto Norbert <norbrt@codeaurora.org>