ar71xx: Fix pll config for qca9558
For qca9558, the pll mii ctrol register should be 0x28 instead of 0x2c.

On our internal board, it used fixed 1000M, and uboot has set the correct value,
so we did not see this issue on ap135.

But on customer board which gmac connect with phy without switch, it may negotiate different speed,
so it can not work normal. Need fix this common issue although only see on customer board,
and set the clock invert bit as 1 by default according to HW requiement.

Change-Id: I1e30b1bdf2005deadad8b41321cc6d276af9c77f
Signed-off-by: xiaofeis <xiaofeis@codeaurora.org>