hostapd: Add vlan_file option to netifd.sh
Other VLAN related options are already being processed in netifd.sh
but the vlan_file option is missing. This option allows the mapping
of vlan IDs to network interfaces and will be used in dynamic VLAN
feature for binding stations to interfaces based on VLAN
assignments. The change is done similarly to the wpa_psk_file
option.

Signed-off-by: Gong Cheng <chengg11@yahoo.com>

Change-Id: Ib350af44105448940526204b46421614d55aaa01
git-svn-id: svn://svn.openwrt.org/openwrt/trunk@46652 3c298f89-4303-0410-b956-a3cf2f4a3e73
Signed-off-by: Peter Oh <poh@codeaurora.org>