[firewall] fixes the SNAT wrong mapping when src_dport is set
when user set up the SNAT with the option src_dport,but use iptable
to check the setting, the iptable rule will becomes dest_port with same
value of the src_dport, this is not expected to the user.

this is because that when dest_port is not set, it will overwrite
it with the src_dport value.

this change will remove the code that assigned the src_dport value to
dest_port value.

Change-Id: Iae44f3b78ebeb1180010dba161b15ee34685ea96
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>