[netifd] shell proto don't handle notify correctly
when enabled ipv6 on the ppp connection, the other apps try to get
the wan's dev interface name, it return the eth0 instead of pppoe-wan
interface name, it will cause other apps won't work.

when ppp is bringing up, the interface name is not set correctly
and it is not brought up by the netifd.

That is because when the shell proto will set the keep to true, in
the netifd, it use this flag to set the new interface name or not,
changed to when it is first time or not bring up, it need update it
the original commit information:
commit d5452cb41927b580df46f3ac81ed9e77a7c83a51
Author: Felix Fietkau <nbd@openwrt.org>
Date:   Thu Apr 10 13:30:54 2014 +0200

	proto-shell: fix a regression with handling multiple proto updates for the same interface

	Preserve the device when keep == true

	Signed-off-by: Felix Fietkau <nbd@openwrt.org>

this patch is generated with command git format-patch and apply it here

Change-Id: I0f879cde6a619e8370d2fab500cf9293286f1a92
Signed-off-by: Felix Fietkau <nbd@openwrt.org>