ipq806x: fix reset to factory settings feature in Dakota boards
When daemons are killed, procd will respawn it if respawn value of that daemon is
set and upgrade_running value is disabled. ubus call will set the upgrade_running
value which tells procd not to respawn any daemons when it is killed. Hence
rootfs_data partition is not in use by any process and can be erased. mkfs.ext4
has been used to format mmc device which is in ext4 format.

Change-Id: I28e632e29015192a535b44c0aaa3bc2bba052a58
Signed-off-by: Parameshwaran Kanagaraj <pkanag@codeaurora.org>