tools: Add new coccinelle package
Coccinelle is used, in particular to generate the backport. As using
the chaos calmer backports package requires to use the chaos calmer
coccinelle package, we'll facilitate the deployment by building it
from OpenWrt rather than requiring each individual to forward-port
the package for its distribution.

This change requires a new tool to be installed on the machine however,
ocaml-nox which is present in most distributions.

Change-Id: I2482c93df08547b1b862fdf93be8b6c6b58799c3
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>