[package] iwinfo: Fix for recursive dependency error
Currently when we issue "make package/symlinks" command in
coconut, we are seeing the below recursive error:
 "tmp/.config-package.in:22124:error: recursive dependency detected!
  symbol PACKAGE_luci-app-statistics depends on PACKAGE_collectd-mod-iwinfo
  symbol PACKAGE_collectd-mod-iwinfo is selected by PACKAGE_luci-app-statistics"

These errors seems to be result of inaccurate dependency in
libiwinfo package, qca-wifi-* package's dependency was part
of DEPENDS as well as PKG_BUILD_DEPENDS, which was making
qsdk build system confused.

Inorder to fix these errors we will remove the unnecessary
dependency and we will add as build dependency only, to
make sure that any dependent package(qca-wifi-*) will be built
first in case it is enabled.

Change-Id: I81ebc894130d3877d0a0ecc10cd17baf181721b3
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>