wireless-tools: Allow for greater than 512 private ioctls
iwpriv queries the list of private ioctls from the kernel and grows its
buffer exponentially if the kernel does not give a size hint. It has a
maximum size of 1000 entries. However, since the starting size is 16
entries, if there are more than 512 private ioctls provided by the
driver, the next attempted size will be 1024. Since the while loop had a
limit of 1000, this number of entries is never tried.

Fix this by allowing for the size to reach 1024 (inclusive).

Change-Id: Ide87f273e5c4b940212029248bc0f8a796509718
Signed-off-by: PavanKumar N <pnande@codeaurora.org>