qca-nss-drv: don't try to load firmware just install symlink
procd still tries to load firmware at boot even if a hotplug
script is installed. It will even try to start loading an
empty file it seems, which can cause a valid firmware file
being loaded to be abruptly stopped.

Let's just have the hotplug script install a symlink to the
correct firmware and let procd handle the loading.

Change-Id: Ib1df4faee0fbed6a9264656fa5c5e1424bc2f2b3
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>