[qca-nss-ecm] Make ECM variant for premium profile
The premium variant has additional dependencies on other kernel modules
that may not be relevant or desired for the basic profile. In this
variant, pass down a make variable to enable the appropriate features.

Update the profiles so that the open profile does not depend on the
premium variant and the premium one does.

Change-Id: I9a86b3da4b7d59715cf10c566889dd64e04ad48f
Signed-off-by: Kiran Kumar C.S.K <kkumarcs@codeaurora.org>