ar71xx: Add 128k size threshold in rootfs size
Currently when a package is enabled, it will reduce the free
space in 16M and 8M board, which does not allow to store user
configs in overlay partition.

In order to fix this problem, we will add additional 128k size
threshold which will keep atleast 128k free space in the image
and will allow user to save its configs.

Change-Id: Iea6ca67d9e14f3a452336fe5f781a6115a49becd
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>