ag71xx: Fix eth0 ping interrupted while down/up eth1 on AP147
On AP147 board, ifconfig down/up eth1 will lead eth0 ping interrupted.
The reason is when ifconfig up eth1 brings up eth1, it will reset all
the switch phys including the phy-4 which is used by eth0 polling, so
eth0 will also down and up, this is root cause.

This change will check the polling mask value from the configure data,
and will not reset the phy which is set in the mask. This phy should be
controlled by the linux polling task.

Change-Id: I84610b2fe08b5a6050ce12d3e0264587c5fa8bba
Signed-off-by: xiaofeis <xiaofeis@codeaurora.org>