scripts: Add dtb name instead of config id in itb image
ITB image was having one generic kernel and several dtb files
and several configurations based on the config id's, so u-boot
was taking the appropriate configurations according to board's
config id.
This change will remove the hardcoded config id's support and
instead of that it will generate the configurations based on the
board names. u-boot will choose the configuration according to
the board name.

Change-Id: I8ce02a9321eaad1625e98d3c120871948fec825e
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>