musl: Add format attribute to some function declarations
GCC and Clang are able to check the format arguments given to a
function and warn the user if there is a error in the format arguments
or if there is a potential uncontrolled format string security problem
in the code. GCC does this automatically for some functions like
printf(), but it is also possible to annotate other functions in a way
that it will check them too. This feature is used by glibc for many
functions. This patch adds the attribute to the some functions of musl
expect for these functions where gcc automatically adds it.

GCC automatically adds checks for these functions: printf, fprintf,
sprintf, scanf, fscanf, sscanf, strftime, vprintf, vfprintf and
vsprintf.

The documentation from gcc is here:
https://gcc.gnu.org/onlinedocs/gcc/Common-Function-Attributes.html

The documentation from Clang is here:
http://clang.llvm.org/docs/AttributeReference.html#format-gnu-format

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@48467 3c298f89-4303-0410-b956-a3cf2f4a3e73
(cherry picked from commit c23c38080482173e3137d8428019ada88b5af11a)

Change-Id: I45456a02ae6b2796f4a9fb1d89017bf43080acd5
Signed-off-by: Anto Norbert <norbrt@codeaurora.org>