hostapd: Update OpenSSL 0.9.8 patch for EAP-FAST support
The 0.9.9 branch was for development purposes only, so no one should be
using that in production and there is not much point in maintaining the
obsolete patch here either. Similarly, the old 0.9.8 versions are
obsolete at this point in time and taken into account the recent OpenSSL
vulnerabilities, anything older than 0.9.8za should not really be used.

Prepare an updated version of the TLS session ticket patch based on the
current OpenSSL 0.9.8za release and remove all the older TLS extension
patches.

[This change is directly cherry picked from the upsteam hostapd commit
http://w1.fi/cgit/hostap/commit/?id=f5fa824e9a86940835e30a5a0b1fd3d8a7c4c640 ]

Change-Id: I8aa1448233d15466a7211ce6273219975e8dbbf4
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>