hostapd: Avoid busy loop in scan result mismatch corner cases
It was possible for interworking_find_network_match() to find a possible
BSS match in a case where more thorough checks in
wpa_supplicant_select_bss() reject network. This itself is fine, in
general, but when combined with wpa_supplicant_fast_associate()
optimization and auto_interworking=1, this resulted in a busy loop of up
to five seconds and a possible stack overflow due to recursion in that
loop.

Fix this by limiting the Interworking wpa_supplicant_fast_associate()
call to be used only once per scan iteration, so that new scan
operations can be completed before going through the scan results again.

[This change is directly cherry picked from the upsteam hostapd commit
http://w1.fi/cgit/hostap/commit/?id=a8826b1848003eb84a6a287d4d71183ba2192dd6 ]

Change-Id: If6b260270cdb27176d8304178fe418a0c9980471
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>