hostapd: Fix payload length validation for Commit and Confirm
The length of the received Commit and Confirm message payloads was not
checked before reading them. This could result in a buffer read
overflow when processing an invalid message.

Fix this by verifying that the payload is of expected length before
processing it. In addition, enforce correct state transition sequence to
make sure there is no unexpected behavior if receiving a Commit/Confirm
message before the previous exchanges have been completed.

Thanks to Kostya Kortchinsky of Google security team for discovering and
reporting this issue.

[This change is directly cherry picked from the upsteam hostapd commit
http://w1.fi/cgit/hostap/commit/?id=dd2f043c9c43d156494e33d7ce22db96e6ef42c7 ]

Change-Id: I88341beb76800758980f29c72af7b550da597634
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>