hostapd: EAP-pwd server: Fix payload length validation for Commit and Confirm
The length of the received Commit and Confirm message payloads was not
checked before reading them. This could result in a buffer read
overflow when processing an invalid message.

Fix this by verifying that the payload is of expected length before
processing it. In addition, enforce correct state transition sequence to
make sure there is no unexpected behavior if receiving a Commit/Confirm
message before the previous exchanges have been completed.

Thanks to Kostya Kortchinsky of Google security team for discovering and
reporting this issue.

[This change is directly cherry picked from the upsteam hostapd commit
http://w1.fi/cgit/hostap/commit/?id=e28a58be26184c2a23f80b410e0997ef1bd5d578 ]

Change-Id: I35bed681ea56188c5c6d985fa69ecf10fc16223f
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>