iwinfo: Avoid stack smashing detected by increasing buffer size
iwinfo command gives stack smashing detected issue when the
interface name is more than 10 characters.

When iwinfo command is given without any parameters, It searches
for all the interfaces in /sys/class/net/ and we have interface
name called "bonding_masters" in that path which is more than
10 characters and this gives the stack smashing detected issue.
So increased the buffer length.

Change-Id: Id0f1925e38304eeda8ae653003273525f3d13be4
Signed-off-by: Anto Norbert <norbrt@codeaurora.org>