netfilter.mk: Use config depending on the linux version.
In Linux-3.17, xt_nat module is moved to CONFIG_NETFILTER_XT_NAT and
iptable_nat module is moved to CONFIG_IP_NF_NAT. While using
linux-3.14, these modules are not installed in lib/modules/ because
the CONFIG name in linux and netfilter.mk doesn't match. So,
CONFIG name for these modules are used based on the linux version.

Change-Id: Iec96994bbc5601ef861f41512cb00d872a482e30
Signed-off-by: Sachin Sundar <ssundar@codeaurora.org>