ar71xx: Add more board data changes based on available driver.
CR:1018018 FR:34235
The Radio names are changing with new driver. So changing
the preinit script as well.
Due to driver modularization FR34235, the OL device comes up
before the DA device. This makes the name of the internal DA
device dynamic (wifi0 or wifi1)
So based on if the PCI device (i.e OL device) is available
we create the calibration file accordingly.

Change-Id: I72aa45404dfaff418e4dffdd191153e45e12c527
Signed-off-by: Venkata Krishna Sundararajan <vsundara@codeaurora.org>
Signed-off-by: Vinay Adella <vadella@codeaurora.org>