remove luci-app-qos and replace with its dependencies
the luci-app-qos and qos-scripts package conflicts with other
QoS management systems which need to manipulate iptables and
qdiscs as well.

after some internal discussion, the parts that are actually
needed from qos-scripts are iptables and tc commands which
enable testing of the QoS subsystem.  the scripts that
manage iptables and qdiscs are not actually required and its
the scripts that conflict other QoS management systems.

Change-Id: I23d9da1b94505bf2bfc79f8124e558603e8be7dc
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>