ipq806x: Fix saving network config after sysupgrade in nand.
On sysupgrade in nand, the /etc/config changes are not preserved.
After sysupgrade completes and reboots, it is considered as first_boot
and overlay is deleted. So, instead of untar the config tarball before
reboot, copy it into the overlay. While mounting root during the preinit,
the untar of the tarball will be done and the configs are preserved.

Change-Id: I3a7856ad6b044a7812a655b144a1324900886123
Signed-off-by: Sachin Sundar <ssundar@codeaurora.org>