[netifd] bonding: add option to force enslaving order
The enslaving order is important for FST to run properly.
The second enslaved interface will get the MAC address of the first one
and only the 11ad driver supports MAC address change so need to ensure
the enslaving order as defined in the uci configuration.
This functionality is controlled by uci command with the entry of:
network.bond.enforce_slaves_order.

CRs-Fixed: 1026404

Change-Id: I18e7a8d433d42abc984f1cac43ee05fafe6470fb
Signed-off-by: Alexei Avshalom Lazar <ailizaro@codeaurora.org>