include: Add ubi volume names as argument in UbinizeImage
Currently UbinizeImage takes the hardcoded volume names in image
as "ubi_rootfs" and "rootfs_data". Since we are adding linux-4.4
in CC, where rootfs volume name is "rootfs".

To make it more flexible we removed the hardcoded name of ubi volume
and we will make them as argument to UbinizeImage.

Change-Id: I7cfac165e02360036a7b7703f29dd3444fd84935
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>