include: Add PKG_SOURCE_BRANCH support package definition
When we request to mirror some git project to CAF, after
some times when the project moves to some other location
or if it becomes dead, then we request CAF to mirror the
same project from another active git location.

Since now there are two mirror locations for the same
project, CAF will make a separate branch for each link,
This is good idea from CAF perspective that we have all
the mirrors request as different branches in same project

So when we clone this type of project, normally git will
be confused for HEAD and will throw warning message which
causes git clone failuere

git clone failure warning/error:
 "warning: remote HEAD refers to nonexistent ref, unable to checkout"

To avoid this problem, we will add branch name support in
package makefile, so this will help us to clone the project
from CAF mirror.

Change-Id: I8a86b2a6cffcdfc8d61533d917d7df6a5ca944e3
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>