base-files: Add support to skip the file in sysupgrade backup files
Currently sysupgrade will take all the changed files in backup,
will create the sysupgrade.tgz to restore the configurations.
Sometimes we may need to skip some file to be retained during
sysupgrade.

In order to achieve this, we added FILTER_FILES which will have
the list of files which we want to be skiped from backup.

Change-Id: I138e593176f1b28f88b2bc5f04fb5f42a89c4bf3
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>