hostapd: Stop AP beaconing before removing interfaces
For drivers like ath10k which receives SWBA/Rx packets based
on events/messages from the driver, it is better to do expicitly
do stop AP(send vdev down to firmware) and then remove the interface
from cfg80211/ mac80211. This avoids a race conditions of packets being
received in the driver in a small window where the interface is deleted
in mac80211, while driver's remove interface is not yet called. This
should help in addressing some of the call traces/ warning messages
in the rx hot path and SWBA event handling

Change-Id: I9cf72f955f8f0eb06beae92c98731cb77b34c671
Signed-off-by: Mohammed Shafi Shajakhan <mohammed@codeaurora.org>
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>