[qca-nss-ecm] USe the correct debugfs paths.
The conntrack notifier was moved from frontend directories
to the ECM's top directory. Conntrack notifier is
common for the frontends anymore. So, instead of checking
the conntrack notifier directory, now we check the frontend
directories (ecm_nss_ipv4 or ecm_sfe_ipv4) to decide if the
underlying acceleration engine is NSS or SFE.

The frontend stop files are also moved out of the frontend
dirctories. They are now common for both NSS and SFE. So,
in the init script, now we are using the common path
for NSS and SFE.

Change-Id: Id23022ace5215edd5b07edcea6b866a727a8df79
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>