ar71xx: kexec_load - Append board serial number & current date/time
There is request to append board serial number and current date/time to CRASH
DUMP file(s) to distinguish them when multiple DUTs are connected to same tftp
server.

To utilize this feature, user has to create BOOTARGS serialnumber. This has to
be unique string to distinguish DUMP for all DUTs connected to same tftp server.

Change-Id: I5c9e7f0cdebc6072b78f6a101d41a65b4af017fc
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>