ar71xx: kexec_load - Append board serial number & current date/time
There is request to append board serial number and current date/time to CRASH
DUMP file(s) to distinguish them when multiple DUTs are connected to same tftp
server.
To utilize this feature, user has to create BOOTARGS SerialNumber. This has to
be unique string to distinguish DUMP for all DUTs connected to same tftp server.

Change-Id: Ib36a84686212b6e40ecddfdb987f0d6f6e5c57fa
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>