ipq806x: Format rootfs_data in emmc sysupgrade
In emmc sysupgrade with configuration preservation, rootfs_data was not
erased. So, apart from the config files, the whole of rootfs_data is
saved which is not the case in nor and nand sysupgrades. To make
all sysupgrade same, erase rootfs_data in emmc sysupgrade even in case of
configuration preservation.

Change-Id: I09bb70c5c4b96785ff8075a80314a3f5c69d007c
Signed-off-by: Sachin Sundar <ssundar@codeaurora.org>