[base-files] Add generic debug mechanism
The goal of this mechanism is to provide an easy and decentralized way
for a package to register a set of debugging commands, in the case where
someone wants to gather all the logs in the system and store them for
later analysis.

The applications can register a set of debug commands by adding a
script in /lib/debug. The "sysdebug" command will recursively execute
all these scripts, log their output and format it accordingly.

Change-Id: Ia57df5e37244f2626c2279610fd4f8ef95e29e83
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Signed-off-by: Anto Norbert <norbrt@codeaurora.org>