openvpn: add extra respawn parameters.
This change protects the openvpn instances to be marked as "in a crash
loop" and thereby the connection retries will run infinitely.

When the remote site of an openvpn connection goes down for some time
(network failure etc.) the openvpn instance in an openwrt/lede device
should not stop retrying to establish the connection.

With the current limit of 5 retries, there is a user interaction
required, which isn't really what you want when the device should
simply do everything to keep the vpn connection up.

Signed-off-by: Martin Schiller <ms@dev.tdt.de>
(cherry picked from commit 06c49dbccf193219b094cbee54fde0048e410eec)

Change-Id: I8a8c19848c6fcf572571787649c7bfa81fedd7a7
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>