libs/toolchain: Modify libc shared library spec
Current spec is not working for external toolchain and following libraries are
missing in libc package
ld-uClibc.so.0,
ld-uClibc.so.1 &
libuClibc-1.0.14.so.

Due to this, libc package will not include these essential libraries and all
executable binaries will STOP working.

Fixed in such a way, so that all these libraries will get included.

Change-Id: Iebc5a3a62a3ede019e8184a10d7b293ab079660b
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>