elfutils: Drop linux-kernel-module.c dwf1 support
Currently linux-3.18 and linux-4.4 has support for
finding the given address from offline dwarfs instead
of online kernel dwarfs.
We added this support in linux-3.14 too, now our all
linux versions are based on offline dwarfs, so we do
not need linux-kernel-modules.c to provide the dwf1
functions support in elfutils and removing it.

Change-Id: I331555096b29beb9143401470e7d31ce515ab5ff
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>