qsdk: Add Signed kernel module support
Kernel modules are  to be signed when Kernel module signing
option CONFIG_MODULE_SIG is enabled along with the automatic
sign option CONFIG_MODULE_SIG_ALL.
When enabled, the Linux kernel will only load kernel modules
that are digitally signed with the proper key. This allows
further hardening of the system by disallowing unsigned kernel
modules, or kernel modules signed with the wrong key, to be loaded.

Change-Id: I7aa320eeafa1cd4d87fb399d46971907c3bc055c
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>