ipq806x: Use statically linked upgraded procd daemon.
After linux-3.18, sysupgrade was not able save configs as overlay was
mounted due to high reference count. This is because the procd daemon is
dynamically linked to the libraries. So, upgraded binary which is statically
linked. This new daemon will kill procd and all other process and and invoke
a new sysupgrade process.

Change-Id: I8208d4dd78c50a4747507d9a6ec4669a0fb08158
Signed-off-by: Sachin Sundar <ssundar@codeaurora.org>