ppp: propagate master firewall zone to dynamic slave interface
Assign the virtual DHCPv6 interface the firewall zone of the parent interface
so fw3 knows the zone to which the virtual DHCPv6 interface belongs.
This guarantees the firewall settings are applied correctly for the virtual
DHCPv6 interface and allows to query the zone to which the virtual DHCPv6
interface belongs via the fw3 network option.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>
(cherry picked from commit 21f25bc4a325b3075d6426047640e9cf6176894c)

Change-Id: I76b93bc07d5fa904f23b4a2dbb996b44f891afe3
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>