ipq807x_64: Enable EXT4_FS alone in eMMC boot support.
We have eMMC boot with overlay support for EXT4_FS only, so
disable the other eMMC filesystem types and keep only ext4.

VFAT_FS is enabled(=m) from openwrt kernel modules, so we
should not keep it as inbuilt.

OVERLAY_FS is enabled from generic/config-4.4, so there is
no need to enable or disable it from SoC specific config.

Change-Id: Ieb18c29a9f634590df309b5e1bc1e0a697c1768f
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>