Config-kernel.in: Add support to enable/disable DEBUG_INFO_REDUCED
CONFIG_DEBUG_INFO_REDUCED is enabled by default when CONFIG_DEBUG_INFO
is enabled, which makes some of the kernel symbols unavailable in core
dumps and it reduces the debugging info in image.
In order to enable this for normal builds and disable for debug builds,
we can add this in OpenWrt configs Config-kernel.in, which gives us a
flexibility to enable/disable this based on profile/builds(normal or
debug build).

Change-Id: I5b413d119c03a01b47183ee30192876297a151c4
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>