Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>
Git-commit: cb5132bb35698cc0c743e34fe0e845dfc4c3e410
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: Ic8fbb6149379b6ce34ba1793d2be1c8758ab6022
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>