Fix PTK rekeying to generate a new ANonce
The Authenticator state machine path for PTK rekeying ended up bypassing
the AUTHENTICATION2 state where a new ANonce is generated when going
directly to the PTKSTART state since there is no need to try to
determine the PMK again in such a case. This is far from ideal since the
new PTK would depend on a new nonce only from the supplicant.

Fix this by generating a new ANonce when moving to the PTKSTART state
for the purpose of starting new 4-way handshake to rekey PTK.

Signed-off-by: Jouni Malinen <j@w1.fi>

Git-commit: 0adc9b28b39d414d5febfff752f6a1576f785c85
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: Ic37eff8548c2ea8a85ecdcb365568b2fd8192151
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>