TDLS: Reject TPK-TK reconfiguration
Do not try to reconfigure the same TPK-TK to the driver after it has
been successfully configured. This is an explicit check to avoid issues
related to resetting the TX/RX packet number. There was already a check
for this for TPK-M2 (retries of that message are ignored completely), so
that behavior does not get modified.

For TPK-M3, the TPK-TK could have been reconfigured, but that was
followed by immediate teardown of the link due to an issue in updating
the STA entry. Furthermore, for TDLS with any real security (i.e.,
ignoring open/WEP), the TPK message exchange is protected on the AP path
and simple replay attacks are not feasible.

As an additional corner case, make sure the local nonce gets updated if
the peer uses a very unlikely "random nonce" of all zeros.

Signed-off-by: Jouni Malinen <j@w1.fi>

Git-commit: ff89af96e5a35c86f50330d2b86c18323318a60c
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I85500d3bf7ee9b53649043014cbcc600bca83ec0
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>