Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>

Git-commit: cb5132bb35698cc0c743e34fe0e845dfc4c3e41
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: Iffc0ea9c88138efb62c909e1de9c477b8c6154b8
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>