ipq807x: sysupgrade: Add 2s delay before reboot in sysupgrade
During sysupgrade, before rebooting, the serial console buffer
has to be flushed, which needs some delay. As the reboot happens
before it is being flushed, some of the console prints are missing.

There are shutdown and flush functions defined in the serial driver,
which does the flush of the remaining console buffers in the queue.
But these functions are not being called when ‘reboot’ or ‘reboot -f’
command is given during the sysupgrade path. This is causing the issue.

Let’s say even if the flush happens from our serial driver, the system
might reboot before the buffers are being flushed in the upper layers
i.e. UART layer and TTY layer. The buffer may still be stuck in one
of the upper layers and the print might not come on the console.
There are flush operations defined in each of these layers, but
the system might reboot before the prints come on the console.

A piece of analysis,
The msm_shutdown() function, which does the flush operations and close
the serial communication, is called when reboot command is given
directly in the console. But the function is not called when reboot
command is given during the sysupgrade path. The reboot command
differs in behavior.

So, added 2s delay before reboot, for the flush to happen in
all the layers.

Change-Id: I1a643fca36fa1330d1df5890cc688602311c4b23
Signed-off-by: Balaji Jagadeesan <bjagadee@codeaurora.org>